update doubleEq, factor out setup hooks
