+++ suite/shared/src/main/scala/org/hammerlab/Suite.scala
@@ -1,65 +1,51 @@
 package org.hammerlab
 
 import cats.Eq
-import cats.kernel.instances.ListInstances2
+import org.hammerlab.math.syntax.FuzzyCmp.FuzzyCmpOps
 import org.hammerlab.math.syntax.Tolerance
 import org.hammerlab.test.CanEq.withConversion
-import org.hammerlab.test.{ CanEq, MkEqDerivation }
-import org.scalatest.{ BeforeAndAfterAll, BeforeAndAfterEach, FunSuite, Matchers }
+import org.hammerlab.test.{ Afters, Befores, CanEq, MkEqDerivation }
+import org.scalatest.{ FunSuite, Matchers }
 
-import scala.collection.mutable.ArrayBuffer
+import scala.math.abs
 
 abstract class Suite
   extends FunSuite
     with Matchers
-    with BeforeAndAfterAll
-    with BeforeAndAfterEach
-    with MkEqDerivation
-    with ListInstances2 {
+    with Befores
+    with Afters
+    with MkEqDerivation {
 
   implicit val  intOrder = cats.instances. int.catsKernelStdOrderForInt
   implicit val longOrder = cats.instances.long.catsKernelStdOrderForLong
 
-  implicit var ε = Tolerance(1e-6)
+  /** Fuzziness for [[Double]] assertions / equality-comparisons; see [[doubleEq]] */
+  implicit var ε: Tolerance = 1e-6
 
+  /** Convenience-setter for [[ε]] */
   def tolerance(d: Double): Unit = {
     ε = d
   }
 
+  /** Default [[CanEq]] instances for [[Double]]s vs. [[Int]]s */
   implicit val int2double : CanEq[Double, Int] = withConversion[Double, Int]
   implicit val int2long   : CanEq[  Long, Int] = withConversion[  Long, Int]
 
-  implicit def doubleEq(implicit tolerance: Tolerance): Eq[Double] =
+  /**
+   * Consider [[Double]]s to be equal if their ratio if their ratio differs from [[1]] by less than or equal to [[ε]],
+   * or if one of them equals zero and the other's absolute value is less than or equal to [[ε]]
+   */
+  implicit def doubleEq: Eq[Double] =
     new Eq[Double] {
       override def eqv(x: Double, y: Double) =
-        x + tolerance >= y &&
-        y + tolerance >= x
+        if (x == 0 || y == 0)
+          abs(x + y) + 1 <= ε
+        else
+          new FuzzyCmpOps(x).===(y)
     }
 
   def ===[T, U](t1: T, t2: U)(implicit canEqual: CanEq[T, U]): Unit = {
     if (!canEqual.eqv(t1, t2))
       fail(s"$t1 didn't match $t2")
-  }
-
-  private val befores = ArrayBuffer[() ⇒ Unit]()
-
-  def before(fn: ⇒ Unit): Unit = {
-    befores += (() ⇒ fn)
-  }
-
-  final override def beforeEach(): Unit = {
-    super.beforeEach()
-    for { beforeFn ← befores } { beforeFn() }
-  }
-
-  private val afters = ArrayBuffer[() ⇒ Unit]()
-
-  def after(fn: ⇒ Unit): Unit = {
-    afters += (() ⇒ fn)
-  }
-
-  final override def afterEach(): Unit = {
-    super.afterEach()
-    for { afterFn ← afters } { afterFn() }
   }
 }
