+++ suite/shared/src/main/scala/org/hammerlab/test/Befores.scala
@@ -1,0 +1,22 @@
+package org.hammerlab.test
+
+import org.hammerlab.Suite
+import org.scalatest.{ BeforeAndAfterAll, BeforeAndAfterEach }
+
+import scala.collection.mutable.ArrayBuffer
+
+trait Befores
+  extends BeforeAndAfterAll
+    with BeforeAndAfterEach {
+  self: Suite ⇒
+  private val befores = ArrayBuffer[() ⇒ Unit]()
+
+  def before(fn: ⇒ Unit): Unit = {
+    befores += (() ⇒ fn)
+  }
+
+  final override def beforeEach(): Unit = {
+    super.beforeEach()
+    for { beforeFn ← befores } { beforeFn() }
+  }
+}
